DDF-2773 Make commons source spec for other source specs to extends - Rename *Test files to *Spec for consistency
