+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/csw/discover/DiscoverCswSourceSpecSpec.groovy
@@ -13,22 +13,16 @@
  **/
 package org.codice.ddf.admin.sources.csw.discover
 
-import org.apache.cxf.jaxrs.client.WebClient
 import org.codice.ddf.admin.api.fields.FunctionField
 import org.codice.ddf.admin.common.fields.common.HostField
 import org.codice.ddf.admin.common.report.message.DefaultMessages
 import org.codice.ddf.admin.sources.csw.CswSourceUtils
 import org.codice.ddf.admin.sources.fields.CswProfile
 import org.codice.ddf.admin.sources.fields.type.CswSourceConfigurationField
-import org.codice.ddf.cxf.SecureCxfClientFactory
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import spock.lang.Shared
-import spock.lang.Specification
 
-import javax.ws.rs.core.Response
-
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
-
-class DiscoverCswSourceTest extends Specification {
+class DiscoverCswSourceSpecSpec extends SourceCommonsSpec {
 
     @Shared
             ddfCswResponse = this.getClass().getClassLoader().getResource('responses/csw/DDFCswGetCapabilities.xml').text
@@ -273,19 +267,5 @@ class DiscoverCswSourceTest extends Specification {
         def cswUtils = new CswSourceUtils()
         cswUtils.setRequestUtils(requestUtils)
         return cswUtils
-    }
-
-    def createMockFactory(int statusCode, String responseBody) {
-        def mockResponse = Mock(Response)
-        mockResponse.getStatus() >> statusCode
-        mockResponse.readEntity(String.class) >> responseBody
-
-        def mockWebClient = Mock(WebClient)
-        mockWebClient.get() >> mockResponse
-
-        def mockFactory = Mock(SecureCxfClientFactory)
-        mockFactory.getClient() >> mockWebClient
-
-        return mockFactory
     }
 }
