+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/opensearch/persist/CreateOpenSearchConfigurationSpec.groovy
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.admin.sources.wfs.persist
+package org.codice.ddf.admin.sources.opensearch.persist
 
 import ddf.catalog.source.FederatedSource
 import org.codice.ddf.admin.api.fields.FunctionField
@@ -20,35 +20,27 @@ import org.codice.ddf.admin.configurator.Configurator
 import org.codice.ddf.admin.configurator.ConfiguratorFactory
 import org.codice.ddf.admin.configurator.OperationReport
 import org.codice.ddf.admin.sources.SourceMessages
-import org.codice.ddf.admin.sources.fields.WfsVersion
-import org.codice.ddf.admin.sources.fields.type.WfsSourceConfigurationField
+import org.codice.ddf.admin.sources.fields.type.OpenSearchSourceConfigurationField
+import org.codice.ddf.admin.sources.fields.type.SourceConfigField
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import org.codice.ddf.internal.admin.configurator.actions.FeatureActions
 import org.codice.ddf.internal.admin.configurator.actions.ManagedServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceReader
-import spock.lang.Specification
 
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
+class CreateOpenSearchConfigurationSpec extends SourceCommonsSpec {
 
-class CreateWfsConfigurationTest extends Specification {
-
-    static RESULT_ARGUMENT_PATH = [CreateWfsConfiguration.FIELD_NAME]
+    static RESULT_ARGUMENT_PATH = [CreateOpenSearchConfiguration.FIELD_NAME]
 
     static BASE_PATH = [RESULT_ARGUMENT_PATH, FunctionField.ARGUMENT].flatten()
 
-    static CONFIG_PATH = [BASE_PATH, WfsSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
+    static CONFIG_PATH = [BASE_PATH, OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
 
-    static SOURCE_NAME_PATH = [CONFIG_PATH, SOURCE_NAME].flatten()
+    static SOURCE_NAME_PATH = [CONFIG_PATH, SourceConfigField.SOURCE_NAME_FIELD_NAME].flatten()
 
     static ENDPOINT_URL_PATH = [CONFIG_PATH, ENDPOINT_URL].flatten()
 
-    static WFS_VERSION = WfsVersion.DEFAULT_FIELD_NAME;
-
-    static WFS_VERSION_PATH = [CONFIG_PATH, WFS_VERSION].flatten()
-
-    static TEST_WFS_VERSION = WfsVersion.Wfs1.WFS_VERSION_1
-
-    CreateWfsConfiguration createWfsConfiguration
+    CreateOpenSearchConfiguration createOpenSearchConfiguration
 
     ConfiguratorFactory configuratorFactory
 
@@ -56,11 +48,11 @@ class CreateWfsConfigurationTest extends Specification {
 
     ServiceReader serviceReader
 
-    Configurator configurator
+    FeatureActions featureActions
 
     ManagedServiceActions managedServiceActions
 
-    FeatureActions featureActions
+    Configurator configurator
 
     FederatedSource federatedSource
 
@@ -76,35 +68,33 @@ class CreateWfsConfigurationTest extends Specification {
         federatedSource = Mock(FederatedSource)
         federatedSource.getId() >> TEST_SOURCENAME
         federatedSources.add(federatedSource)
-        configuratorFactory = Mock(ConfiguratorFactory) {
-            getConfigurator() >> configurator
-        }
+        configuratorFactory = Mock(ConfiguratorFactory)
+        configuratorFactory.getConfigurator() >> configurator
 
-        createWfsConfiguration = new CreateWfsConfiguration(configuratorFactory, serviceActions,
+        createOpenSearchConfiguration = new CreateOpenSearchConfiguration(configuratorFactory, serviceActions,
                 managedServiceActions, serviceReader, featureActions)
     }
 
-    def 'Successfully create new WFS configuration'() {
+    def 'Successfully create new OpenSearch configuration'() {
         setup:
-        createWfsConfiguration.setValue(createWfsArgs())
+        createOpenSearchConfiguration.setValue(createFunctionArgs())
         serviceReader.getServices(_, _) >> []
         configurator.commit(_, _) >> mockReport(false)
 
         when:
-        def report = createWfsConfiguration.getValue()
+        def report = createOpenSearchConfiguration.getValue()
 
         then:
-        report.result() != null
         report.result().getValue()
     }
 
-    def 'Fail to create new WFS config due to duplicate source name'() {
+    def 'Fail to create new OpenSearch config due to duplicate source name'() {
         setup:
-        createWfsConfiguration.setValue(createWfsArgs())
+        createOpenSearchConfiguration.setValue(createFunctionArgs())
         serviceReader.getServices(_, _) >> federatedSources
 
         when:
-        def report = createWfsConfiguration.getValue()
+        def report = createOpenSearchConfiguration.getValue()
 
         then:
         report.result() == null
@@ -113,28 +103,28 @@ class CreateWfsConfigurationTest extends Specification {
         report.messages().get(0).path == SOURCE_NAME_PATH
     }
 
-    def 'Fail to create new WFS config due to failure to commit'() {
+    def 'Fail to create new OpenSearch config due to failure to commit'() {
         setup:
-        createWfsConfiguration.setValue(createWfsArgs())
+        createOpenSearchConfiguration.setValue(createFunctionArgs())
         serviceReader.getServices(_, _) >> []
 
         when:
-        def report = createWfsConfiguration.getValue()
+        def report = createOpenSearchConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(false)
         1 * configurator.commit(_, _) >> mockReport(true)
         !report.result().getValue()
         report.messages().size() == 1
-        report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
         report.messages().get(0).path == RESULT_ARGUMENT_PATH
+        report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Return false when wfs feature fails to start'() {
+    def 'Return false when opensearch feature fails to start'() {
         when:
-        createWfsConfiguration.setValue(createWfsArgs())
+        createOpenSearchConfiguration.setValue(createFunctionArgs())
         serviceReader.getServices(_, _) >> []
-        def report = createWfsConfiguration.getValue()
+        def report = createOpenSearchConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(true)
@@ -144,29 +134,28 @@ class CreateWfsConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Fail due to missing required fields'() {
+    def 'Fail when missing required fields'() {
         when:
-        def report = createWfsConfiguration.getValue()
+        def report = createOpenSearchConfiguration.getValue()
 
         then:
         report.result() == null
-        report.messages().size() == 3
+        report.messages().size() == 2
         report.messages().count {
             it.getCode() == DefaultMessages.MISSING_REQUIRED_FIELD
-        } == 3
-        report.messages()*.getPath() == [SOURCE_NAME_PATH, ENDPOINT_URL_PATH, WFS_VERSION_PATH]
-    }
-
-    def createWfsArgs() {
-        def config = new WfsSourceConfigurationField().wfsVersion(TEST_WFS_VERSION)
-                .endpointUrl('https://localhost:8993/geoserver/wfs').sourceName(TEST_SOURCENAME)
-        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
-        return [(WfsSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
+        } == 2
+        report.messages()*.getPath() == [SOURCE_NAME_PATH, ENDPOINT_URL_PATH]
     }
 
     def mockReport(boolean hasError) {
         def report = Mock(OperationReport)
         report.containsFailedResults() >> hasError
         return report
+    }
+
+    def createFunctionArgs() {
+        def config = new OpenSearchSourceConfigurationField().endpointUrl('https://localhost:8993').sourceName(TEST_SOURCENAME)
+        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
+        return [(OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
     }
 }
