+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/opensearch/persist/UpdateOpenSearchConfigurationSpec.groovy
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.admin.sources.wfs.persist
+package org.codice.ddf.admin.sources.opensearch.persist
 
 import ddf.catalog.source.FederatedSource
 import org.codice.ddf.admin.api.fields.FunctionField
@@ -20,37 +20,29 @@ import org.codice.ddf.admin.configurator.Configurator
 import org.codice.ddf.admin.configurator.ConfiguratorFactory
 import org.codice.ddf.admin.configurator.OperationReport
 import org.codice.ddf.admin.sources.SourceMessages
-import org.codice.ddf.admin.sources.fields.WfsVersion
-import org.codice.ddf.admin.sources.fields.type.WfsSourceConfigurationField
+import org.codice.ddf.admin.sources.fields.type.OpenSearchSourceConfigurationField
+import org.codice.ddf.admin.sources.fields.type.SourceConfigField
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import org.codice.ddf.internal.admin.configurator.actions.FeatureActions
 import org.codice.ddf.internal.admin.configurator.actions.ManagedServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceReader
-import spock.lang.Specification
 
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
+class UpdateOpenSearchConfigurationSpec extends SourceCommonsSpec {
 
-class UpdateWfsConfigurationTest extends Specification {
-
-    static RESULT_ARGUMENT_PATH = [UpdateWfsConfiguration.FIELD_NAME]
+    static RESULT_ARGUMENT_PATH = [UpdateOpenSearchConfiguration.FIELD_NAME]
 
     static BASE_PATH = [RESULT_ARGUMENT_PATH, FunctionField.ARGUMENT].flatten()
 
-    static CONFIG_PATH = [BASE_PATH, WfsSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
+    static CONFIG_PATH = [BASE_PATH, OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
 
-    static SOURCE_NAME_PATH = [CONFIG_PATH, SOURCE_NAME].flatten()
+    static SOURCE_NAME_PATH = [CONFIG_PATH, SourceConfigField.SOURCE_NAME_FIELD_NAME].flatten()
 
     static ENDPOINT_URL_PATH = [CONFIG_PATH, ENDPOINT_URL].flatten()
 
     static PID_PATH = [CONFIG_PATH, PID].flatten()
 
-    static WFS_VERSION = WfsVersion.DEFAULT_FIELD_NAME;
-
-    static WFS_VERSION_PATH = [CONFIG_PATH, WFS_VERSION].flatten()
-
-    static TEST_WFS_VERSION = WfsVersion.Wfs1.WFS_VERSION_1
-
-    UpdateWfsConfiguration updateWfsConfiguration
+    UpdateOpenSearchConfiguration updateOpenSearchConfiguration
 
     ConfiguratorFactory configuratorFactory
 
@@ -58,11 +50,11 @@ class UpdateWfsConfigurationTest extends Specification {
 
     ServiceReader serviceReader
 
-    Configurator configurator
+    FeatureActions featureActions
 
     ManagedServiceActions managedServiceActions
 
-    FeatureActions featureActions
+    Configurator configurator
 
     FederatedSource federatedSource
 
@@ -78,53 +70,52 @@ class UpdateWfsConfigurationTest extends Specification {
         federatedSource = Mock(FederatedSource)
         federatedSource.getId() >> TEST_SOURCENAME
         federatedSources.add(federatedSource)
-        configuratorFactory = Mock(ConfiguratorFactory) {
-            getConfigurator() >> configurator
-        }
+        configuratorFactory = Mock(ConfiguratorFactory)
+        configuratorFactory.getConfigurator() >> configurator
 
-        updateWfsConfiguration = new UpdateWfsConfiguration(configuratorFactory, serviceActions,
+        updateOpenSearchConfiguration = new UpdateOpenSearchConfiguration(configuratorFactory, serviceActions,
                 managedServiceActions, serviceReader, featureActions)
     }
 
-    def 'Successfully update WFS configuration'() {
+    def 'Successfully update existing OpenSearch configuration'() {
         setup:
-        updateWfsConfiguration.setValue(createWfsUpdateArgs())
-        serviceActions.read(_ as String) >> [(ID): TEST_SOURCENAME]
+        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
+        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
         serviceReader.getServices(_, _) >> []
         configurator.commit(_, _) >> mockReport(false)
 
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         report.result().getValue()
     }
 
-    def 'Fail to update due to existing source name specified by pid'() {
+    def 'Fail to update config due to existing source name'() {
         setup:
-        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
         serviceActions.read(_) >> [(ID): 'updatedName']
-        serviceReader.getServices(_, _) >> [new TestSource(S_PID, 'updatedName', false),
-                                            new TestSource("existingSource", TEST_SOURCENAME, false)]
+        serviceReader.getServices(_, _) >> [new TestSource(S_PID, 'updatedName', false), new TestSource("existingSource", TEST_SOURCENAME, false)]
 
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         report.result() == null
         report.messages().size() == 1
-        report.messages().get(0).code == SourceMessages.DUPLICATE_SOURCE_NAME
         report.messages().get(0).path == SOURCE_NAME_PATH
+        report.messages().get(0).code == SourceMessages.DUPLICATE_SOURCE_NAME
     }
 
-    def 'Fail configuration update due to failure to commit'() {
+    def 'Fail to update config due to failure to commit'() {
         setup:
-        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
         serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
         serviceReader.getServices(_, _) >> []
+        configurator.commit(_, _) >> mockReport(true)
 
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(false)
@@ -135,13 +126,13 @@ class UpdateWfsConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Fail to update WFS config due to no existing config specified by pid'() {
+    def 'Fail to update config due to no existing source specified by the pid'() {
         setup:
-        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
         serviceActions.read(S_PID) >> [:]
 
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         report.result() == null
@@ -150,14 +141,14 @@ class UpdateWfsConfigurationTest extends Specification {
         report.messages().get(0).path == RESULT_ARGUMENT_PATH
     }
 
-    def 'Return false when wfs feature fails to start'() {
+    def 'Return false when opensearch feature fails to start'() {
         setup:
-        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
         serviceReader.getServices(_, _) >> []
         serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
 
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(true)
@@ -167,30 +158,29 @@ class UpdateWfsConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Fail due to missing required fields'() {
+    def 'Fail when missing required fields'() {
         when:
-        def report = updateWfsConfiguration.getValue()
+        def report = updateOpenSearchConfiguration.getValue()
 
         then:
         report.result() == null
-        report.messages().size() == 4
+        report.messages().size() == 3
         report.messages().count {
             it.getCode() == DefaultMessages.MISSING_REQUIRED_FIELD
-        } == 4
-        report.messages()*.getPath() == [PID_PATH, SOURCE_NAME_PATH, ENDPOINT_URL_PATH, WFS_VERSION_PATH]
-    }
-
-    def createWfsUpdateArgs() {
-        def config = new WfsSourceConfigurationField().wfsVersion(TEST_WFS_VERSION)
-                .endpointUrl('https://localhost:8993/geoserver/wfs').sourceName(TEST_SOURCENAME)
-                .pid(S_PID)
-        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
-        return [(WfsSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
+        } == 3
+        report.messages()*.getPath() == [PID_PATH, SOURCE_NAME_PATH, ENDPOINT_URL_PATH]
     }
 
     def mockReport(boolean hasError) {
         def report = Mock(OperationReport)
         report.containsFailedResults() >> hasError
         return report
+    }
+
+    def createUpdateFunctionArgs() {
+        def config = new OpenSearchSourceConfigurationField().endpointUrl('https://localhost:8993').sourceName(TEST_SOURCENAME)
+                .pid(S_PID)
+        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
+        return [(OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
     }
 }
