+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/wfs/persist/CreateWfsConfigurationSpec.groovy
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.admin.sources.opensearch.persist
+package org.codice.ddf.admin.sources.wfs.persist
 
 import ddf.catalog.source.FederatedSource
 import org.codice.ddf.admin.api.fields.FunctionField
@@ -20,31 +20,33 @@ import org.codice.ddf.admin.configurator.Configurator
 import org.codice.ddf.admin.configurator.ConfiguratorFactory
 import org.codice.ddf.admin.configurator.OperationReport
 import org.codice.ddf.admin.sources.SourceMessages
-import org.codice.ddf.admin.sources.fields.type.OpenSearchSourceConfigurationField
-import org.codice.ddf.admin.sources.fields.type.SourceConfigField
+import org.codice.ddf.admin.sources.fields.WfsVersion
+import org.codice.ddf.admin.sources.fields.type.WfsSourceConfigurationField
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import org.codice.ddf.internal.admin.configurator.actions.FeatureActions
 import org.codice.ddf.internal.admin.configurator.actions.ManagedServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceReader
-import spock.lang.Specification
 
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
+class CreateWfsConfigurationSpec extends SourceCommonsSpec {
 
-class UpdateOpenSearchConfigurationTest extends Specification {
-
-    static RESULT_ARGUMENT_PATH = [UpdateOpenSearchConfiguration.FIELD_NAME]
+    static RESULT_ARGUMENT_PATH = [CreateWfsConfiguration.FIELD_NAME]
 
     static BASE_PATH = [RESULT_ARGUMENT_PATH, FunctionField.ARGUMENT].flatten()
 
-    static CONFIG_PATH = [BASE_PATH, OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
+    static CONFIG_PATH = [BASE_PATH, WfsSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
 
-    static SOURCE_NAME_PATH = [CONFIG_PATH, SourceConfigField.SOURCE_NAME_FIELD_NAME].flatten()
+    static SOURCE_NAME_PATH = [CONFIG_PATH, SOURCE_NAME].flatten()
 
     static ENDPOINT_URL_PATH = [CONFIG_PATH, ENDPOINT_URL].flatten()
 
-    static PID_PATH = [CONFIG_PATH, PID].flatten()
+    static WFS_VERSION = WfsVersion.DEFAULT_FIELD_NAME;
 
-    UpdateOpenSearchConfiguration updateOpenSearchConfiguration
+    static WFS_VERSION_PATH = [CONFIG_PATH, WFS_VERSION].flatten()
+
+    static TEST_WFS_VERSION = WfsVersion.Wfs1.WFS_VERSION_1
+
+    CreateWfsConfiguration createWfsConfiguration
 
     ConfiguratorFactory configuratorFactory
 
@@ -52,11 +54,11 @@ class UpdateOpenSearchConfigurationTest extends Specification {
 
     ServiceReader serviceReader
 
-    FeatureActions featureActions
+    Configurator configurator
 
     ManagedServiceActions managedServiceActions
 
-    Configurator configurator
+    FeatureActions featureActions
 
     FederatedSource federatedSource
 
@@ -72,85 +74,65 @@ class UpdateOpenSearchConfigurationTest extends Specification {
         federatedSource = Mock(FederatedSource)
         federatedSource.getId() >> TEST_SOURCENAME
         federatedSources.add(federatedSource)
-        configuratorFactory = Mock(ConfiguratorFactory)
-        configuratorFactory.getConfigurator() >> configurator
+        configuratorFactory = Mock(ConfiguratorFactory) {
+            getConfigurator() >> configurator
+        }
 
-        updateOpenSearchConfiguration = new UpdateOpenSearchConfiguration(configuratorFactory, serviceActions,
+        createWfsConfiguration = new CreateWfsConfiguration(configuratorFactory, serviceActions,
                 managedServiceActions, serviceReader, featureActions)
     }
 
-    def 'Successfully update existing OpenSearch configuration'() {
+    def 'Successfully create new WFS configuration'() {
         setup:
-        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
-        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
+        createWfsConfiguration.setValue(createWfsArgs())
         serviceReader.getServices(_, _) >> []
         configurator.commit(_, _) >> mockReport(false)
 
         when:
-        def report = updateOpenSearchConfiguration.getValue()
+        def report = createWfsConfiguration.getValue()
 
         then:
+        report.result() != null
         report.result().getValue()
     }
 
-    def 'Fail to update config due to existing source name'() {
+    def 'Fail to create new WFS config due to duplicate source name'() {
         setup:
-        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
-        serviceActions.read(_) >> [(ID): 'updatedName']
-        serviceReader.getServices(_, _) >> [new TestSource(S_PID, 'updatedName', false), new TestSource("existingSource", TEST_SOURCENAME, false)]
+        createWfsConfiguration.setValue(createWfsArgs())
+        serviceReader.getServices(_, _) >> federatedSources
 
         when:
-        def report = updateOpenSearchConfiguration.getValue()
+        def report = createWfsConfiguration.getValue()
 
         then:
         report.result() == null
         report.messages().size() == 1
-        report.messages().get(0).path == SOURCE_NAME_PATH
         report.messages().get(0).code == SourceMessages.DUPLICATE_SOURCE_NAME
+        report.messages().get(0).path == SOURCE_NAME_PATH
     }
 
-    def 'Fail to update config due to failure to commit'() {
+    def 'Fail to create new WFS config due to failure to commit'() {
         setup:
-        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
-        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
+        createWfsConfiguration.setValue(createWfsArgs())
         serviceReader.getServices(_, _) >> []
-        configurator.commit(_, _) >> mockReport(true)
 
         when:
-        def report = updateOpenSearchConfiguration.getValue()
+        def report = createWfsConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(false)
         1 * configurator.commit(_, _) >> mockReport(true)
         !report.result().getValue()
         report.messages().size() == 1
-        report.messages().get(0).path == RESULT_ARGUMENT_PATH
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
-    }
-
-    def 'Fail to update config due to no existing source specified by the pid'() {
-        setup:
-        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
-        serviceActions.read(S_PID) >> [:]
-
-        when:
-        def report = updateOpenSearchConfiguration.getValue()
-
-        then:
-        report.result() == null
-        report.messages().size() == 1
-        report.messages().get(0).code == DefaultMessages.NO_EXISTING_CONFIG
         report.messages().get(0).path == RESULT_ARGUMENT_PATH
     }
 
-    def 'Return false when opensearch feature fails to start'() {
-        setup:
-        updateOpenSearchConfiguration.setValue(createUpdateFunctionArgs())
-        serviceReader.getServices(_, _) >> []
-        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
-
+    def 'Return false when wfs feature fails to start'() {
         when:
-        def report = updateOpenSearchConfiguration.getValue()
+        createWfsConfiguration.setValue(createWfsArgs())
+        serviceReader.getServices(_, _) >> []
+        def report = createWfsConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(true)
@@ -160,9 +142,9 @@ class UpdateOpenSearchConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Fail when missing required fields'() {
+    def 'Fail due to missing required fields'() {
         when:
-        def report = updateOpenSearchConfiguration.getValue()
+        def report = createWfsConfiguration.getValue()
 
         then:
         report.result() == null
@@ -170,19 +152,19 @@ class UpdateOpenSearchConfigurationTest extends Specification {
         report.messages().count {
             it.getCode() == DefaultMessages.MISSING_REQUIRED_FIELD
         } == 3
-        report.messages()*.getPath() == [PID_PATH, SOURCE_NAME_PATH, ENDPOINT_URL_PATH]
+        report.messages()*.getPath() == [SOURCE_NAME_PATH, ENDPOINT_URL_PATH, WFS_VERSION_PATH]
+    }
+
+    def createWfsArgs() {
+        def config = new WfsSourceConfigurationField().wfsVersion(TEST_WFS_VERSION)
+                .endpointUrl('https://localhost:8993/geoserver/wfs').sourceName(TEST_SOURCENAME)
+        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
+        return [(WfsSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
     }
 
     def mockReport(boolean hasError) {
         def report = Mock(OperationReport)
         report.containsFailedResults() >> hasError
         return report
-    }
-
-    def createUpdateFunctionArgs() {
-        def config = new OpenSearchSourceConfigurationField().endpointUrl('https://localhost:8993').sourceName(TEST_SOURCENAME)
-                .pid(S_PID)
-        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
-        return [(OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME) : config.getValue()]
     }
 }
