+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/wfs/persist/DeleteWfsConfigurationSpec.groovy
@@ -10,102 +10,96 @@
  * Lesser General Public License for more details. A copy of the GNU Lesser General Public License
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
- */
-package org.codice.ddf.admin.sources.csw.persist
+ **/
+package org.codice.ddf.admin.sources.wfs.persist
 
 import org.codice.ddf.admin.api.fields.FunctionField
 import org.codice.ddf.admin.common.report.message.DefaultMessages
 import org.codice.ddf.admin.configurator.Configurator
 import org.codice.ddf.admin.configurator.ConfiguratorFactory
 import org.codice.ddf.admin.configurator.OperationReport
-import org.codice.ddf.admin.sources.services.CswServiceProperties
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import org.codice.ddf.internal.admin.configurator.actions.ManagedServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceActions
-import spock.lang.Specification
 
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
+class DeleteWfsConfigurationSpec extends SourceCommonsSpec {
 
-class DeleteCswConfigurationTest extends Specification {
-
-    DeleteCswConfiguration deleteCswConfiguration
+    DeleteWfsConfiguration deleteWfsConfiguration
 
     ConfiguratorFactory configuratorFactory
 
     Configurator configurator
 
-    ServiceActions serviceActions
+    private ServiceActions serviceActions
 
-    static TEST_CSW_URL = "testCswUrl"
+    static RESULT_ARGUMENT_PATH = [DeleteWfsConfiguration.FIELD_NAME]
 
-    static EVENT_SERVICE_ADDRESS = "eventServiceAddress"
+    static BASE_PATH = [RESULT_ARGUMENT_PATH, FunctionField.ARGUMENT].flatten()
 
-    static TEST_EVENT_SERVICE_ADDRESS = "testEventServiceAddress"
-
-    static RESULT_ARGUMENT_PATH = [DeleteCswConfiguration.FIELD_NAME]
-
-    static BASE_PATH = [DeleteCswConfiguration.FIELD_NAME, FunctionField.ARGUMENT]
-
-    static SERVICE_PID_PATH = [BASE_PATH, PID].flatten()
+    static PID_PATH = [BASE_PATH, PID].flatten()
 
     def functionArgs = [
             (PID): S_PID
     ]
 
-    def configToDelete = createCswConfigToDelete()
-
     def setup() {
         configurator = Mock(Configurator)
+        configuratorFactory = Mock(ConfiguratorFactory) {
+            getConfigurator() >> getConfigurator()
+        }
         serviceActions = Mock(ServiceActions)
         def managedServiceActions = Mock(ManagedServiceActions)
 
-        configuratorFactory = Mock(ConfiguratorFactory) {
-            getConfigurator() >> configurator
-        }
-        deleteCswConfiguration = new DeleteCswConfiguration(configuratorFactory, this.serviceActions, managedServiceActions)
+        deleteWfsConfiguration = new DeleteWfsConfiguration(configuratorFactory, serviceActions,
+                managedServiceActions)
     }
 
-    def 'Successfully deleting CSW configuration returns true'() {
-        when:
-        serviceActions.read(S_PID) >> configToDelete
+    def 'Successfully delete WFS configuration'() {
+        setup:
+        serviceActions.read(S_PID) >> configToBeDeleted
         configurator.commit(_, _) >> mockReport(false)
-        deleteCswConfiguration.setValue(functionArgs)
-        def report = deleteCswConfiguration.getValue()
+        deleteWfsConfiguration.setValue(functionArgs)
+
+        when:
+        def report = deleteWfsConfiguration.getValue()
 
         then:
         report.result() != null
         report.result().getValue()
     }
 
-    def 'Fail with no existing config found with provided pid'() {
+    def 'Fail to discover WFS config when no existing config found with provided pid'() {
+        setup:
+        serviceActions.read(S_PID) >> [:]
+        deleteWfsConfiguration.setValue(functionArgs)
+
         when:
-        serviceActions.read(_ as String) >> [:]
-        deleteCswConfiguration.setValue(functionArgs)
-        def report = deleteCswConfiguration.getValue()
+        def report = deleteWfsConfiguration.getValue()
 
         then:
         report.result() == null
         report.messages().size() == 1
-        report.messages().get(0).path == RESULT_ARGUMENT_PATH
         report.messages().get(0).code == DefaultMessages.NO_EXISTING_CONFIG
+        report.messages().get(0).path == RESULT_ARGUMENT_PATH
     }
 
-    def 'Error while committing deleted configuration with the given servicePid'() {
+    def 'Error while committing delete configuration with given pid'() {
         when:
-        serviceActions.read(S_PID) >> configToDelete
+        serviceActions.read(S_PID) >> configToBeDeleted
         configurator.commit(_, _) >> mockReport(true)
-        deleteCswConfiguration.setValue(functionArgs)
-        def report = deleteCswConfiguration.getValue()
+        deleteWfsConfiguration.setValue(functionArgs)
+        def report = deleteWfsConfiguration.getValue()
 
         then:
         !report.result().getValue()
         report.messages().size() == 1
-        report.messages().get(0).path == RESULT_ARGUMENT_PATH
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
+        report.messages().get(0).path == RESULT_ARGUMENT_PATH
     }
 
     def 'Fail when missing required fields'() {
         when:
-        def report = deleteCswConfiguration.getValue()
+        def report = deleteWfsConfiguration.getValue()
 
         then:
         report.result() == null
@@ -113,19 +107,12 @@ class DeleteCswConfigurationTest extends Specification {
         report.messages().count {
             it.getCode() == DefaultMessages.MISSING_REQUIRED_FIELD
         } == 1
-        report.messages()*.getPath() == [SERVICE_PID_PATH]
+        report.messages()*.getPath() == [PID_PATH]
     }
 
     private def mockReport(boolean hasError) {
         def report = Mock(OperationReport)
         report.containsFailedResults() >> hasError
         return report
-    }
-
-    private def createCswConfigToDelete() {
-        configToDelete = configToBeDeleted
-        configToDelete.put(EVENT_SERVICE_ADDRESS, TEST_EVENT_SERVICE_ADDRESS)
-        configToDelete.put(CswServiceProperties.CSW_URL, TEST_CSW_URL)
-        return configToDelete;
     }
 }
