+++ query/sources/impl/src/test/groovy/org/codice/ddf/admin/sources/wfs/persist/UpdateWfsConfigurationSpec.groovy
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.admin.sources.opensearch.persist
+package org.codice.ddf.admin.sources.wfs.persist
 
 import ddf.catalog.source.FederatedSource
 import org.codice.ddf.admin.api.fields.FunctionField
@@ -20,29 +20,35 @@ import org.codice.ddf.admin.configurator.Configurator
 import org.codice.ddf.admin.configurator.ConfiguratorFactory
 import org.codice.ddf.admin.configurator.OperationReport
 import org.codice.ddf.admin.sources.SourceMessages
-import org.codice.ddf.admin.sources.fields.type.OpenSearchSourceConfigurationField
-import org.codice.ddf.admin.sources.fields.type.SourceConfigField
+import org.codice.ddf.admin.sources.fields.WfsVersion
+import org.codice.ddf.admin.sources.fields.type.WfsSourceConfigurationField
+import org.codice.ddf.admin.sources.test.SourceCommonsSpec
 import org.codice.ddf.internal.admin.configurator.actions.FeatureActions
 import org.codice.ddf.internal.admin.configurator.actions.ManagedServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceActions
 import org.codice.ddf.internal.admin.configurator.actions.ServiceReader
-import spock.lang.Specification
 
-import static org.codice.ddf.admin.sources.test.SourceTestCommons.*
+class UpdateWfsConfigurationSpec extends SourceCommonsSpec {
 
-class CreateOpenSearchConfigurationTest extends Specification {
-
-    static RESULT_ARGUMENT_PATH = [CreateOpenSearchConfiguration.FIELD_NAME]
+    static RESULT_ARGUMENT_PATH = [UpdateWfsConfiguration.FIELD_NAME]
 
     static BASE_PATH = [RESULT_ARGUMENT_PATH, FunctionField.ARGUMENT].flatten()
 
-    static CONFIG_PATH = [BASE_PATH, OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
+    static CONFIG_PATH = [BASE_PATH, WfsSourceConfigurationField.DEFAULT_FIELD_NAME].flatten()
 
-    static SOURCE_NAME_PATH = [CONFIG_PATH, SourceConfigField.SOURCE_NAME_FIELD_NAME].flatten()
+    static SOURCE_NAME_PATH = [CONFIG_PATH, SOURCE_NAME].flatten()
 
     static ENDPOINT_URL_PATH = [CONFIG_PATH, ENDPOINT_URL].flatten()
 
-    CreateOpenSearchConfiguration createOpenSearchConfiguration
+    static PID_PATH = [CONFIG_PATH, PID].flatten()
+
+    static WFS_VERSION = WfsVersion.DEFAULT_FIELD_NAME;
+
+    static WFS_VERSION_PATH = [CONFIG_PATH, WFS_VERSION].flatten()
+
+    static TEST_WFS_VERSION = WfsVersion.Wfs1.WFS_VERSION_1
+
+    UpdateWfsConfiguration updateWfsConfiguration
 
     ConfiguratorFactory configuratorFactory
 
@@ -50,11 +56,11 @@ class CreateOpenSearchConfigurationTest extends Specification {
 
     ServiceReader serviceReader
 
-    FeatureActions featureActions
+    Configurator configurator
 
     ManagedServiceActions managedServiceActions
 
-    Configurator configurator
+    FeatureActions featureActions
 
     FederatedSource federatedSource
 
@@ -70,33 +76,37 @@ class CreateOpenSearchConfigurationTest extends Specification {
         federatedSource = Mock(FederatedSource)
         federatedSource.getId() >> TEST_SOURCENAME
         federatedSources.add(federatedSource)
-        configuratorFactory = Mock(ConfiguratorFactory)
-        configuratorFactory.getConfigurator() >> configurator
+        configuratorFactory = Mock(ConfiguratorFactory) {
+            getConfigurator() >> configurator
+        }
 
-        createOpenSearchConfiguration = new CreateOpenSearchConfiguration(configuratorFactory, serviceActions,
+        updateWfsConfiguration = new UpdateWfsConfiguration(configuratorFactory, serviceActions,
                 managedServiceActions, serviceReader, featureActions)
     }
 
-    def 'Successfully create new OpenSearch configuration'() {
+    def 'Successfully update WFS configuration'() {
         setup:
-        createOpenSearchConfiguration.setValue(createFunctionArgs())
+        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        serviceActions.read(_ as String) >> [(ID): TEST_SOURCENAME]
         serviceReader.getServices(_, _) >> []
         configurator.commit(_, _) >> mockReport(false)
 
         when:
-        def report = createOpenSearchConfiguration.getValue()
+        def report = updateWfsConfiguration.getValue()
 
         then:
         report.result().getValue()
     }
 
-    def 'Fail to create new OpenSearch config due to duplicate source name'() {
+    def 'Fail to update due to existing source name specified by pid'() {
         setup:
-        createOpenSearchConfiguration.setValue(createFunctionArgs())
-        serviceReader.getServices(_, _) >> federatedSources
+        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        serviceActions.read(_) >> [(ID): 'updatedName']
+        serviceReader.getServices(_, _) >> [new TestSource(S_PID, 'updatedName', false),
+                                            new TestSource("existingSource", TEST_SOURCENAME, false)]
 
         when:
-        def report = createOpenSearchConfiguration.getValue()
+        def report = updateWfsConfiguration.getValue()
 
         then:
         report.result() == null
@@ -105,13 +115,14 @@ class CreateOpenSearchConfigurationTest extends Specification {
         report.messages().get(0).path == SOURCE_NAME_PATH
     }
 
-    def 'Fail to create new OpenSearch config due to failure to commit'() {
+    def 'Fail configuration update due to failure to commit'() {
         setup:
-        createOpenSearchConfiguration.setValue(createFunctionArgs())
+        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
         serviceReader.getServices(_, _) >> []
 
         when:
-        def report = createOpenSearchConfiguration.getValue()
+        def report = updateWfsConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(false)
@@ -122,11 +133,29 @@ class CreateOpenSearchConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Return false when opensearch feature fails to start'() {
+    def 'Fail to update WFS config due to no existing config specified by pid'() {
+        setup:
+        updateWfsConfiguration.setValue(createWfsUpdateArgs())
+        serviceActions.read(S_PID) >> [:]
+
         when:
-        createOpenSearchConfiguration.setValue(createFunctionArgs())
+        def report = updateWfsConfiguration.getValue()
+
+        then:
+        report.result() == null
+        report.messages().size() == 1
+        report.messages().get(0).code == DefaultMessages.NO_EXISTING_CONFIG
+        report.messages().get(0).path == RESULT_ARGUMENT_PATH
+    }
+
+    def 'Return false when wfs feature fails to start'() {
+        setup:
+        updateWfsConfiguration.setValue(createWfsUpdateArgs())
         serviceReader.getServices(_, _) >> []
-        def report = createOpenSearchConfiguration.getValue()
+        serviceActions.read(_) >> [(ID): TEST_SOURCENAME]
+
+        when:
+        def report = updateWfsConfiguration.getValue()
 
         then:
         1 * configurator.commit(_, _) >> mockReport(true)
@@ -136,28 +165,30 @@ class CreateOpenSearchConfigurationTest extends Specification {
         report.messages().get(0).code == DefaultMessages.FAILED_PERSIST
     }
 
-    def 'Fail when missing required fields'() {
+    def 'Fail due to missing required fields'() {
         when:
-        def report = createOpenSearchConfiguration.getValue()
+        def report = updateWfsConfiguration.getValue()
 
         then:
         report.result() == null
-        report.messages().size() == 2
+        report.messages().size() == 4
         report.messages().count {
             it.getCode() == DefaultMessages.MISSING_REQUIRED_FIELD
-        } == 2
-        report.messages()*.getPath() == [SOURCE_NAME_PATH, ENDPOINT_URL_PATH]
+        } == 4
+        report.messages()*.getPath() == [PID_PATH, SOURCE_NAME_PATH, ENDPOINT_URL_PATH, WFS_VERSION_PATH]
+    }
+
+    def createWfsUpdateArgs() {
+        def config = new WfsSourceConfigurationField().wfsVersion(TEST_WFS_VERSION)
+                .endpointUrl('https://localhost:8993/geoserver/wfs').sourceName(TEST_SOURCENAME)
+                .pid(S_PID)
+        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
+        return [(WfsSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
     }
 
     def mockReport(boolean hasError) {
         def report = Mock(OperationReport)
         report.containsFailedResults() >> hasError
         return report
-    }
-
-    def createFunctionArgs() {
-        def config = new OpenSearchSourceConfigurationField().endpointUrl('https://localhost:8993').sourceName(TEST_SOURCENAME)
-        config.credentials().username(TEST_USERNAME).password(TEST_PASSWORD)
-        return [(OpenSearchSourceConfigurationField.DEFAULT_FIELD_NAME): config.getValue()]
     }
 }
